Move lox.Token -> ast.Token
